Additional fixes.
