+++ sway-core/src/semantic_analysis/ast_node/declaration/function.rs
@@ -157,25 +157,12 @@ impl TypedFunctionDeclaration {
         // insert parameters and generic type declarations into namespace
         let mut namespace = namespace.clone();
 
-        // insert type parameters as Unknown types
-        let type_mapping = insert_type_parameters(&type_parameters);
-
         // update the types in the type parameters, insert the type parameters
         // into the decl namespace, and check to see if the type parameters
         // shadow one another
         for type_parameter in type_parameters.iter_mut() {
             check!(
-                type_parameter.update_types_with_self(&type_mapping, &mut namespace, self_type),
-                return err(warnings, errors),
-                warnings,
-                errors
-            );
-            let type_parameter_decl = TypedDeclaration::GenericTypeForFunctionScope {
-                name: type_parameter.name_ident.clone(),
-                type_id: type_parameter.type_id,
-            };
-            check!(
-                namespace.insert_symbol(type_parameter.name_ident.clone(), type_parameter_decl),
+                TypeParameter::type_check(type_parameter, &mut namespace),
                 continue,
                 warnings,
                 errors
@@ -341,9 +328,10 @@ impl TypedFunctionDeclaration {
                 |TypedFunctionParameter {
                      r#type, type_span, ..
                  }| {
-                    resolve_type(*r#type, type_span)
+                    #[allow(clippy::needless_borrow)]
+                    resolve_type(*r#type, &type_span)
                         .expect("unreachable I think?")
-                        .to_selector_name(type_span)
+                        .to_selector_name(&type_span)
                 },
             )
             .filter_map(|name| name.ok(&mut warnings, &mut errors))
