+++ sway-core/src/semantic_analysis/ast_node/mod.rs
@@ -483,47 +483,15 @@ impl TypedAstNode {
                             );
                             decl
                         }
-                        Declaration::AbiDeclaration(AbiDeclaration {
-                            name,
-                            interface_surface,
-                            methods,
-                            span,
-                        }) => {
-                            // type check the interface surface and methods
-                            // We don't want the user to waste resources by contract calling
-                            // themselves, and we don't want to do more work in the compiler,
-                            // so we don't support the case of calling a contract's own interface
-                            // from itself. This is by design.
-                            let self_type_for_interface = insert_type(TypeInfo::SelfType);
-                            let mut new_interface_surface = vec![];
-                            for trait_fn in interface_surface.into_iter() {
-                                new_interface_surface.push(check!(
-                                    TypedTraitFn::type_check(
-                                        trait_fn,
-                                        namespace,
-                                        self_type_for_interface
-                                    ),
-                                    continue,
-                                    warnings,
-                                    errors
-                                ));
-                            }
-
-                            // type check these for errors but don't actually use them yet -- the real
-                            // ones will be type checked with proper symbols when the ABI is implemented
-                            let _methods = check!(
-                                type_check_trait_methods(methods.clone(), namespace, self_type),
-                                vec![],
+                        Declaration::AbiDeclaration(abi_decl) => {
+                            let abi_decl = check!(
+                                TypedAbiDeclaration::type_check(abi_decl, namespace, self_type),
+                                return err(warnings, errors),
                                 warnings,
                                 errors
                             );
-
-                            let decl = TypedDeclaration::AbiDeclaration(TypedAbiDeclaration {
-                                interface_surface: new_interface_surface,
-                                methods,
-                                name: name.clone(),
-                                span,
-                            });
+                            let name = abi_decl.name.clone();
+                            let decl = TypedDeclaration::AbiDeclaration(abi_decl);
                             namespace.insert_symbol(name, decl.clone());
                             decl
                         }
@@ -791,308 +759,6 @@ fn reassignment(
         })
         .map(TypedDeclaration::StorageReassignment),
     }
-}
-
-/// Recursively handle supertraits by adding all their interfaces and methods to some namespace
-/// which is meant to be the namespace of the subtrait in question
-fn handle_supertraits(
-    supertraits: &[Supertrait],
-    trait_namespace: &mut Namespace,
-) -> CompileResult<()> {
-    let mut warnings = Vec::new();
-    let mut errors = Vec::new();
-
-    for supertrait in supertraits.iter() {
-        match trait_namespace
-            .resolve_call_path(&supertrait.name)
-            .ok(&mut warnings, &mut errors)
-            .cloned()
-        {
-            Some(TypedDeclaration::TraitDeclaration(TypedTraitDeclaration {
-                ref interface_surface,
-                ref methods,
-                ref supertraits,
-                ..
-            })) => {
-                // insert dummy versions of the interfaces for all of the supertraits
-                trait_namespace.insert_trait_implementation(
-                    supertrait.name.clone(),
-                    TypeInfo::SelfType,
-                    interface_surface
-                        .iter()
-                        .map(|x| x.to_dummy_func(Mode::NonAbi))
-                        .collect(),
-                );
-
-                // insert dummy versions of the methods of all of the supertraits
-                let dummy_funcs = check!(
-                    convert_trait_methods_to_dummy_funcs(methods, trait_namespace),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                trait_namespace.insert_trait_implementation(
-                    supertrait.name.clone(),
-                    TypeInfo::SelfType,
-                    dummy_funcs,
-                );
-
-                // Recurse to insert dummy versions of interfaces and methods of the *super*
-                // supertraits
-                check!(
-                    handle_supertraits(supertraits, trait_namespace),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-            }
-            Some(TypedDeclaration::AbiDeclaration(_)) => {
-                errors.push(CompileError::AbiAsSupertrait {
-                    span: supertrait.name.span().clone(),
-                })
-            }
-            _ => errors.push(CompileError::TraitNotFound {
-                name: supertrait.name.clone(),
-            }),
-        }
-    }
-
-    ok((), warnings, errors)
-}
-
-fn type_check_trait_methods(
-    methods: Vec<FunctionDeclaration>,
-    namespace: &mut Namespace,
-    self_type: TypeId,
-) -> CompileResult<Vec<TypedFunctionDeclaration>> {
-    let mut warnings = vec![];
-    let mut errors = vec![];
-    let mut methods_buf = Vec::new();
-    for FunctionDeclaration {
-        body,
-        name: fn_name,
-        parameters,
-        span,
-        return_type,
-        type_parameters,
-        return_type_span,
-        purity,
-        ..
-    } in methods
-    {
-        parameters.clone().into_iter().for_each(
-            |FunctionParameter {
-                 name,
-                 type_id: ref r#type,
-                 ..
-             }| {
-                let r#type = check!(
-                    namespace.resolve_type_with_self(
-                        look_up_type_id(*r#type),
-                        insert_type(TypeInfo::SelfType),
-                        &name.span(),
-                        EnforceTypeArguments::Yes
-                    ),
-                    insert_type(TypeInfo::ErrorRecovery),
-                    warnings,
-                    errors,
-                );
-                namespace.insert_symbol(
-                    name.clone(),
-                    TypedDeclaration::VariableDeclaration(TypedVariableDeclaration {
-                        name: name.clone(),
-                        body: TypedExpression {
-                            expression: TypedExpressionVariant::FunctionParameter,
-                            return_type: r#type,
-                            is_constant: IsConstant::No,
-                            span: name.span(),
-                        },
-                        // TODO allow mutable function params?
-                        is_mutable: VariableMutability::Immutable,
-                        const_decl_origin: false,
-                        type_ascription: r#type,
-                    }),
-                );
-            },
-        );
-        // check the generic types in the arguments, make sure they are in
-        // the type scope
-        let mut generic_params_buf_for_error_message = Vec::new();
-        for param in parameters.iter() {
-            if let TypeInfo::Custom { ref name, .. } = look_up_type_id(param.type_id) {
-                generic_params_buf_for_error_message.push(name.to_string());
-            }
-        }
-        let comma_separated_generic_params = generic_params_buf_for_error_message.join(", ");
-        for FunctionParameter {
-            ref type_id, name, ..
-        } in parameters.iter()
-        {
-            let span = name.span().clone();
-            if let TypeInfo::Custom { name, .. } = look_up_type_id(*type_id) {
-                let args_span = parameters.iter().fold(
-                    parameters[0].name.span().clone(),
-                    |acc, FunctionParameter { name, .. }| Span::join(acc, name.span()),
-                );
-                if type_parameters.iter().any(|TypeParameter { type_id, .. }| {
-                    if let TypeInfo::Custom {
-                        name: this_name, ..
-                    } = look_up_type_id(*type_id)
-                    {
-                        this_name == name.clone()
-                    } else {
-                        false
-                    }
-                }) {
-                    errors.push(CompileError::TypeParameterNotInTypeScope {
-                        name: name.clone(),
-                        span: span.clone(),
-                        comma_separated_generic_params: comma_separated_generic_params.clone(),
-                        fn_name: fn_name.clone(),
-                        args: args_span.as_str().to_string(),
-                    });
-                }
-            }
-        }
-        let parameters = parameters
-            .into_iter()
-            .map(
-                |FunctionParameter {
-                     name,
-                     type_id,
-                     type_span,
-                 }| {
-                    TypedFunctionParameter {
-                        name,
-                        r#type: check!(
-                            namespace.resolve_type_with_self(
-                                look_up_type_id(type_id),
-                                crate::type_engine::insert_type(TypeInfo::SelfType),
-                                &type_span,
-                                EnforceTypeArguments::Yes
-                            ),
-                            insert_type(TypeInfo::ErrorRecovery),
-                            warnings,
-                            errors,
-                        ),
-                        type_span,
-                    }
-                },
-            )
-            .collect::<Vec<_>>();
-
-        // TODO check code block implicit return
-        let return_type = check!(
-            namespace.resolve_type_with_self(
-                return_type,
-                self_type,
-                &return_type_span,
-                EnforceTypeArguments::Yes
-            ),
-            insert_type(TypeInfo::ErrorRecovery),
-            warnings,
-            errors,
-        );
-        let (body, _code_block_implicit_return) = check!(
-            TypedCodeBlock::type_check(TypeCheckArguments {
-                checkee: body,
-                namespace,
-                return_type_annotation: return_type,
-                help_text: "Trait method body's return type does not match up with \
-                                         its return type annotation.",
-                self_type,
-                mode: Mode::NonAbi,
-                opts: TCOpts { purity }
-            }),
-            continue,
-            warnings,
-            errors
-        );
-
-        methods_buf.push(TypedFunctionDeclaration {
-            name: fn_name,
-            body,
-            parameters,
-            span,
-            return_type,
-            type_parameters,
-            // For now, any method declared is automatically public.
-            // We can tweak that later if we want.
-            visibility: Visibility::Public,
-            return_type_span,
-            is_contract_call: false,
-            purity,
-        });
-    }
-    ok(methods_buf, warnings, errors)
-}
-
-/// Convert a vector of FunctionDeclarations into a vector of TypedFunctionDeclarations where only
-/// the parameters and the return types are type checked.
-fn convert_trait_methods_to_dummy_funcs(
-    methods: &[FunctionDeclaration],
-    trait_namespace: &mut Namespace,
-) -> CompileResult<Vec<TypedFunctionDeclaration>> {
-    let mut warnings = vec![];
-    let mut errors = vec![];
-    let dummy_funcs = methods
-        .iter()
-        .map(
-            |FunctionDeclaration {
-                 name,
-                 parameters,
-                 return_type,
-                 return_type_span,
-                 ..
-             }| TypedFunctionDeclaration {
-                purity: Default::default(),
-                name: name.clone(),
-                body: TypedCodeBlock { contents: vec![] },
-                parameters: parameters
-                    .iter()
-                    .map(
-                        |FunctionParameter {
-                             name,
-                             type_id,
-                             type_span,
-                         }| TypedFunctionParameter {
-                            name: name.clone(),
-                            r#type: check!(
-                                trait_namespace.resolve_type_with_self(
-                                    look_up_type_id(*type_id),
-                                    insert_type(TypeInfo::SelfType),
-                                    type_span,
-                                    EnforceTypeArguments::Yes
-                                ),
-                                insert_type(TypeInfo::ErrorRecovery),
-                                warnings,
-                                errors,
-                            ),
-                            type_span: type_span.clone(),
-                        },
-                    )
-                    .collect(),
-                span: name.span(),
-                return_type: check!(
-                    trait_namespace.resolve_type_with_self(
-                        return_type.clone(),
-                        insert_type(TypeInfo::SelfType),
-                        return_type_span,
-                        EnforceTypeArguments::Yes
-                    ),
-                    insert_type(TypeInfo::ErrorRecovery),
-                    warnings,
-                    errors,
-                ),
-                return_type_span: return_type_span.clone(),
-                visibility: Visibility::Public,
-                type_parameters: vec![],
-                is_contract_call: false,
-            },
-        )
-        .collect::<Vec<_>>();
-
-    ok(dummy_funcs, warnings, errors)
 }
 
 /// Used to create a stubbed out function when the function fails to compile, preventing cascading
